refactor(app): move internal/mantis → app/mantis
Mantis client moves into the app module. No tests to migrate.
Imports rewritten across cmd/ and bot/.
